+++ build_boxes.sh
@@ -1,25 +1,25 @@
 #!/bin/bash -eux
 
 bundle exec thor packer:build \
-  --atlas_version=1.0.23 \
+  --atlas_version=1.0.24 \
   --os=debian \
   --os_version=7.9.0 \
   --providers=virtualbox,vmware,parallels
 
 bundle exec thor packer:build \
-  --atlas_version=1.0.23 \
+  --atlas_version=1.0.24 \
   --os=debian \
-  --os_version=8.3.0 \
+  --os_version=8.4.0 \
   --providers=virtualbox,vmware,parallels
 
 bundle exec thor packer:build \
-  --atlas_version=1.0.23 \
+  --atlas_version=1.0.24 \
   --os=ubuntu \
   --os_version=14.04.4-server \
   --providers=virtualbox,vmware,parallels
 
 bundle exec thor packer:build \
-  --atlas_version=1.0.23 \
+  --atlas_version=1.0.24 \
   --os=ubuntu \
   --os_version=15.10-server \
   --providers=virtualbox,parallels